refactor: reorganize project structure and update build configuration
- Move source files to new directory structure (app/global, app/cefview, app/resources)
- Update CMakeLists.txt to reflect new file locations
- Update include paths in source files
- Rename project from CefQuickstart to MikoLauncher
- Update readme.md with new project information and features
