+++ .gitignore
@@ -1,5 +1,5 @@
 build
-app/webapp.cpp
+app/resources/webapp.cpp
 **/build/**
 node_modules
 vscode
