+++ app/resources/binaryresourceprovider.hpp
