+++ app/resources/resourceutil.cpp
